Update patch set 2
Patch Set 2:

As Yatin mentioned this is a fix for newton/mitaka. This patch would fix this in rdo-ci but then we'd need to apply a different fix in packstack upstream jobs (once we promote a new repo in newton it'll start failing also). I'd propose to fix it directly in packstack project adding adding --provision-tempest-repo-revision=14.0.0 in scenarios for stable/newton branch. In any case using master for a stable branch is not a good idea IMO.

Patch-set: 2